Menu permission

+++ app/Http/Controllers/MenuController.php
@@ -25,36 +25,48 @@ class MenuController extends BaseController
      * load menu index route
      */
     public function index(){
-        $this->setPageData('Menu', 'Menu', 'fas fa-th-list');
-        return view('menu.index');
+        if(permission('menu-access')){
+            $this->setPageData('Menu', 'Menu', 'fas fa-th-list');
+            return view('menu.index');
+        }else{
+            $this->unauthorized_access_blocked();
+        }
     }
 
     /**
      * get datatable data
      */
     public function get_datatable_data(Request $request){
-        if($request->ajax()){
-            $output = $this->service->get_datatable_data($request);
+        if(permission('menu-access')){
+            if($request->ajax()){
+                $output = $this->service->get_datatable_data($request);
+            }else{
+                $output = ['status' => 'error', 'message' => 'Unauthorized access blocked!'];
+            }
+    
+            return response()->json($output);
         }else{
-            $output = ['status' => 'error', 'message' => 'Unauthorized access blocked!'];
+            $this->unauthorized_access_blocked();
         }
-
-        return response()->json($output);
     }
 
     /**
      * store or update data
      */
     public function store_or_update(MenuFormRequest $request){
-        if($request->ajax()){
-            $result = $this->service->store_or_update_data($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data has been saved successfully!", $data=null, $response_code=200);
+        if(permission('menu-add') || permission('menu-edit')){
+            if($request->ajax()){
+                $result = $this->service->store_or_update_data($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data has been saved successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Unauthorized access blocked!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Unauthorized access blocked!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            $this->unauthorized_access_blocked();
         }
     }
 
@@ -62,15 +74,19 @@ class MenuController extends BaseController
      * edit menu data
      */
     public function edit(Request $request){
-        if($request->ajax()){
-            $data = $this->service->edit($request);
-            if($data->count() > 0){
-                return $this->response_json($status="success", $message=null, $data=$data, $response_code=201);
+        if(permission('menu-edit')){
+            if($request->ajax()){
+                $data = $this->service->edit($request);
+                if($data->count() > 0){
+                    return $this->response_json($status="success", $message=null, $data=$data, $response_code=201);
+                }else{
+                    return $this->response_json($status="error", $message="Data Not Found", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Data Not Found", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            $this->unauthorized_access_blocked();
         }
     }
 
@@ -78,15 +94,19 @@ class MenuController extends BaseController
      * delete menu data
      */
     public function delete(Request $request){
-        if($request->ajax()){
-            $result = $this->service->delete($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+        if(permission('menu-delete')){
+            if($request->ajax()){
+                $result = $this->service->delete($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            $this->unauthorized_access_blocked();
         }
     }
 
@@ -94,15 +114,19 @@ class MenuController extends BaseController
      * bulk delete data from database
      */
     public function bulk_delete(Request $request){
-        if($request->ajax()){
-            $result = $this->service->bulk_delete($request);
-            if($result){
-                return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+        if(permission('menu-bulk-delete')){
+            if($request->ajax()){
+                $result = $this->service->bulk_delete($request);
+                if($result){
+                    return $this->response_json($status="success", $message="Data deleted successfully!", $data=null, $response_code=200);
+                }else{
+                    return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                }
             }else{
-                return $this->response_json($status="error", $message="Failed to delete data!", $data=null, $response_code=204);
+                return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
             }
         }else{
-            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
+            $this->unauthorized_access_blocked();
         }
     }
